requirement.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_gnu
alsa-lib=1.2.8=h166bdaf_0
attr=2.5.1=h166bdaf_1
boost=1.74.0=py37h796e4cb_5
boost-cpp=1.74.0=h75c5d50_8
brotli=1.0.9=h166bdaf_8
brotli-bin=1.0.9=h166bdaf_8
brotlipy=0.7.0=py37h540881e_1004
bzip2=1.0.8=h7f98852_4
ca-certificates=2022.12.7=ha878542_0
cairo=1.16.0=ha61ee94_1014
certifi=2022.12.7=pyhd8ed1ab_0
cffi=1.15.1=py37h43b0acd_1
charset-normalizer=2.1.1=pyhd8ed1ab_0
colorama=0.4.6=pyhd8ed1ab_0
cryptography=38.0.2=py37h5994e8b_1
cudatoolkit=11.3.1=h9edb442_11
cudnn=8.4.1.50=hed8a83a_0
cycler=0.11.0=pyhd8ed1ab_0
datetime=5.1=pypi_0
dbus=1.13.6=h5008d03_3
et_xmlfile=1.1.0=pyhd8ed1ab_0
expat=2.5.0=h27087fc_0
ffmpeg=4.3=hf484d3e_0
fftw=3.3.10=nompi_hc118613_107
fightingcv-attention=1.0.0=pypi_0
font-ttf-dejavu-sans-mono=2.37=hab24e00_0
font-ttf-inconsolata=3.000=h77eed37_0
font-ttf-source-code-pro=2.038=h77eed37_0
font-ttf-ubuntu=0.83=hab24e00_0
fontconfig=2.14.2=h14ed4e7_0
fonts-conda-ecosystem=1=0
fonts-conda-forge=1=0
fonttools=4.38.0=py37h540881e_0
freetype=2.12.1=hca18f0e_1
future=0.18.2=py37h89c1867_5
gettext=0.21.1=h27087fc_0
giflib=5.2.1=h0b41bf4_3
glib=2.74.1=h6239696_1
glib-tools=2.74.1=h6239696_1
gmp=6.2.1=h58526e2_0
gnutls=3.6.13=h85f3911_1
graphite2=1.3.13=h58526e2_1001
gst-plugins-base=1.22.0=h4243ec0_2
gstreamer=1.22.0=h25f0c4b_2
gstreamer-orc=0.4.33=h166bdaf_0
harfbuzz=6.0.0=h8e241bc_0
icu=70.1=h27087fc_0
idna=3.4=pyhd8ed1ab_0
ilearnplus=1.4=pypi_0
intel-openmp=2022.1.0=h9e868ea_3769
jack=1.9.22=h11f4161_0
joblib=1.2.0=pyhd8ed1ab_0
jpeg=9e=h0b41bf4_3
keyutils=1.6.1=h166bdaf_0
kiwisolver=1.4.4=py37h7cecad7_0
krb5=1.20.1=h81ceb04_0
lame=3.100=h166bdaf_1003
lcms2=2.15=hfd0df8a_0
ld_impl_linux-64=2.40=h41732ed_0
lerc=4.0.0=h27087fc_0
libblas=3.9.0=16_linux64_openblas
libbrotlicommon=1.0.9=h166bdaf_8
libbrotlidec=1.0.9=h166bdaf_8
libbrotlienc=1.0.9=h166bdaf_8
libcap=2.67=he9d0100_0
libcblas=3.9.0=16_linux64_openblas
libclang=15.0.7=default_had23c3d_1
libclang13=15.0.7=default_h3e3d535_1
libcups=2.3.3=h36d4200_3
libdb=6.2.32=h9c3ff4c_0
libdeflate=1.17=h0b41bf4_0
libedit=3.1.20191231=he28a2e2_2
libevent=2.1.10=h28343ad_4
libffi=3.4.2=h7f98852_5
libflac=1.4.2=h27087fc_0
libgcc-ng=12.2.0=h65d4601_19
libgcrypt=1.10.1=h166bdaf_0
libgfortran-ng=12.2.0=h69a702a_19
libgfortran5=12.2.0=h337968e_19
libglib=2.74.1=h606061b_1
libgomp=12.2.0=h65d4601_19
libgpg-error=1.46=h620e276_0
libiconv=1.17=h166bdaf_0
liblapack=3.9.0=16_linux64_openblas
libllvm15=15.0.7=hadd5161_1
libnsl=2.0.0=h7f98852_0
libogg=1.3.4=h7f98852_1
libopenblas=0.3.21=pthreads_h78a6416_3
libopus=1.3.1=h7f98852_1
libpng=1.6.39=h753d276_0
libpq=15.2=hb675445_0
libprotobuf=3.20.2=h6239696_0
libsndfile=1.2.0=hb75c966_0
libsqlite=3.40.0=h753d276_0
libstdcxx-ng=12.2.0=h46fd767_19
libsystemd0=253=h8c4010b_1
libtiff=4.5.0=h6adf6a1_2
libtool=2.4.7=h27087fc_0
libudev1=253=h0b41bf4_1
libuuid=2.32.1=h7f98852_1000
libvorbis=1.3.7=h9c3ff4c_0
libwebp=1.2.4=h1daa5a0_1
libwebp-base=1.2.4=h166bdaf_0
libxcb=1.13=h7f98852_1004
libxkbcommon=1.5.0=h79f4944_1
libxml2=2.10.3=hca2bb57_4
libzlib=1.2.13=h166bdaf_4
lightgbm=3.3.5=pypi_0
lz4-c=1.9.4=hcb278e6_0
magma=2.5.4=hc72dce7_4
matplotlib=3.5.3=py37h89c1867_2
matplotlib-base=3.5.3=py37hf395dca_2
mkl=2022.1.0=hc2b9512_224
mpg123=1.31.3=hcb278e6_0
munkres=1.0.7=py_1
mysql-common=8.0.32=ha901b37_1
mysql-libs=8.0.32=hd7da12d_1
nccl=2.14.3.1=h0800d71_0
ncurses=6.3=h27087fc_1
nettle=3.6=he412f7d_0
ninja=1.11.1=h924138e_0
nspr=4.35=h27087fc_0
nss=3.89=he45b914_0
numpy=1.21.5=py37h976b520_1
openh264=2.1.1=h780b84a_0
openpyxl=3.0.10=py37h540881e_1
openssl=3.1.0=h0b41bf4_0
packaging=23.1=pyhd8ed1ab_0
pandas=1.2.5=py37h295c915_0
pcre2=10.40=hc3806b6_0
pillow=9.3.0=py37h6a678d5_2
pip=23.0.1=pyhd8ed1ab_0
pixman=0.40.0=h36c2ea0_0
ply=3.11=py_1
pthread-stubs=0.4=h36c2ea0_1001
pulseaudio=16.1=hcb278e6_3
pulseaudio-client=16.1=h5195f5e_3
pulseaudio-daemon=16.1=ha8d29e2_3
pycairo=1.21.0=py37h0afab05_1
pycparser=2.21=pyhd8ed1ab_0
pyopenssl=23.1.0=pyhd8ed1ab_0
pyparsing=3.0.9=pyhd8ed1ab_0
pyqt=5.15.7=py37hf30b843_1
pyqt5-sip=12.11.0=py37hd23a5d3_1
pysocks=1.7.1=py37h89c1867_5
python=3.7.12=hf930737_100_cpython
python-dateutil=2.8.2=pyhd8ed1ab_0
python_abi=3.7=3_cp37m
pytorch=1.12.1=cuda112py37hfcfbd4c_200
pytorch-mutex=1.0=cuda
pytz=2023.2=pyhd8ed1ab_0
pyyaml=6.0=py37h540881e_4
qdarkstyle=3.1=pypi_0
qt-main=5.15.8=h5d23da1_6
qtpy=2.3.1=pypi_0
rdkit=2020.09.1=py37h0c252aa_0
readline=8.2=h8228510_1
requests=2.28.2=pyhd8ed1ab_0
scikit-learn=1.0.2=py37hf9e9bfc_0
scipy=1.7.3=py37hf2a6cf1_0
seaborn=0.12.2=pypi_0
setuptools=67.6.0=pyhd8ed1ab_0
sip=6.7.2=py37hd23a5d3_0
six=1.16.0=pyh6c4a22f_0
sleef=3.5.1=h9b69904_2
sqlite=3.40.0=h4ff8645_0
threadpoolctl=3.1.0=pyh8a188c0_0
tk=8.6.12=h27826a3_0
toml=0.10.2=pyhd8ed1ab_0
torchaudio=0.12.1=py37_cu113
torchvision=0.13.1=py37_cu113
tornado=6.2=py37h540881e_0
tqdm=4.64.1=pyhd8ed1ab_0
typing-extensions=4.5.0=hd8ed1ab_0
typing_extensions=4.5.0=pyha770c72_0
unicodedata2=14.0.0=py37h540881e_1
urllib3=1.26.15=pyhd8ed1ab_0
wheel=0.40.0=pyhd8ed1ab_0
xcb-util=0.4.0=h166bdaf_0
xcb-util-image=0.4.0=h166bdaf_0
xcb-util-keysyms=0.4.0=h166bdaf_0
xcb-util-renderutil=0.3.9=h166bdaf_0
xcb-util-wm=0.4.1=h166bdaf_0
xgboost=1.6.2=pypi_0
xkeyboard-config=2.38=h0b41bf4_0
xorg-kbproto=1.0.7=h7f98852_1002
xorg-libice=1.0.10=h7f98852_0
xorg-libsm=1.2.3=hd9c2040_1000
xorg-libx11=1.8.4=h0b41bf4_0
xorg-libxau=1.0.9=h7f98852_0
xorg-libxdmcp=1.1.3=h7f98852_0
xorg-libxext=1.3.4=h0b41bf4_2
xorg-libxrender=0.9.10=h7f98852_1003
xorg-renderproto=0.11.1=h7f98852_1002
xorg-xextproto=7.3.0=h0b41bf4_1003
xorg-xproto=7.0.31=h7f98852_1007
xz=5.2.6=h166bdaf_0
yaml=0.2.5=h7f98852_2
zlib=1.2.13=h166bdaf_4
zope-interface=6.0=pypi_0
zstd=1.5.2=h3eb15da_6
